requirements (1).txt
fastapi>=0.110.0
uvicorn>=0.29.0
watchdog>=4.0.0
psutil>=5.9.0
pydantic>=2.0.0

google-generativeai>=0.4.0
google-cloud-storage>=2.14.0
